+++ src/java/qfs-access-pre-9/pom.xml
@@ -1,0 +1,1 @@
+../qfs-access/pom.xml
