+++ src/java/qfs-access-pre-9/src/main/java/com/quantcast/qfs/access/KfsFileAttr.java
@@ -1,0 +1,1 @@
+../../../../../../../../qfs-access/src/main/java/com/quantcast/qfs/access/KfsFileAttr.java
