+++ src/java/qfs-access-pre-9/src/main/java/com/quantcast/qfs/access/KfsInputChannel.java
@@ -1,11 +1,11 @@
 /**
  * $Id$
  *
- * Created 2025/04/20
+ * Created 2007/09/11
  *
- * @author: Mike Ovsiannikov (Quantcast Corporation)
+ * @author: Sriram Rao (Kosmix Corp.)
  *
- * Copyright 2025 Quantcast Corporation. All rights reserved.
+ * Copyright 2008-2012,2016 Quantcast Corporation. All rights reserved.
  * Copyright 2007 Kosmix Corp.
  *
  * This file is part of Kosmos File System (KFS).
@@ -22,30 +22,24 @@
  * implied. See the License for the specific language governing
  * permissions and limitations under the License.
  *
- * \brief Input channel java 9 style cleanup.
+ * \brief An input channel that does buffered I/O.  This is to reduce
+ * the overhead of JNI calls.
  */
 package com.quantcast.qfs.access;
 
-import java.lang.ref.Cleaner;
+/* A byte channel interface with seek support */
+final public class KfsInputChannel extends KfsInputChannelBase {
 
-final public class KfsOutputChannel extends KfsOutputChannelBase {
-
-    final private Cleaner.Cleanable cleanable;
-
-    KfsOutputChannel(KfsAccessBase ka, int fd) {
+    KfsInputChannel(KfsAccessBase ka, int fd) {
         super(ka, fd);
-        cleanable = registerCleanup();
     }
 
-    private Cleaner.Cleanable registerCleanup() {
-        return KfsAccess.registerCleanup(this, state);
-    }
-
-    public synchronized void close() throws IOException {
+    @Override
+    protected void finalize() throws Throwable {
         try {
-            super.close();
+            state.release();
         } finally {
-            cleanable.clean();
+            super.finalize();
         }
     }
 }
